pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "srt_translator"
version = "0.1.0"
description = "A tool for translating SRT subtitle files."
authors = [
    { name = "YWC Tech" }
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.10"
keywords = ["srt", "translator", "subtitles"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "aisuite[ollama, openai]",
    "python-dotenv",
    "srt"
]

[project.urls]
Homepage = "https://github.com/unclefomotw/srt-translator"
Source = "https://github.com/unclefomotw/srt-translator"
Tracker = "https://github.com/unclefomotw/srt-translator/issues"

[project.scripts]
srt-translate = "srt_translator.cli:main"

[tool.setuptools]
packages = ["srt_translator"]
package-dir = {'' = 'src'}
